refactor row.js to use register_id/callback pattern, move drag/drop out of view.js and into separate file, change dragdrop handlers to underscore_naming
