+++ packages/perspective-viewer/src/html/computed_column.html
@@ -16,7 +16,7 @@ the Apache License 2.0.  The full license can be found in the LICENSE file.
         <div class="psp-cc__container" style="margin-top:-7px;">
             <div class="psp-cc__content">
                 <div id="psp-cc-computation__type"></div>
-                <span contentEditable=true type="text" required maxlength="25" size="10" autocomplete="off" id="psp-cc-name" ondragover="disallowDrop(event)"></span>
+                <span contentEditable=true type="text" required maxlength="25" size="10" autocomplete="off" id="psp-cc-name"></span>
             </div>
             <div class="psp-cc__content psp-cc__content--nomargin">
                 <span class="psp-cc__label psp-cc__error" id="psp-cc__error--name"></span>
@@ -30,7 +30,7 @@ the Apache License 2.0.  The full license can be found in the LICENSE file.
         </div>
         <div class="psp-cc__container" style="margin-top:-12px;">
             <div id="psp-cc-computation-inputs">
-                <!--<div class="psp-cc-computation__input-column" drop-target ondragenter="dragEnter(event)"></div>-->
+                <!--<div class="psp-cc-computation__input-column" drop-target></div>-->
             </div>
         </div>
     </div>
