+++ packages/perspective-viewer/src/less/default.less
@@ -77,7 +77,7 @@
 }
 
 .selected_indicator() {
-    font-family: var(--interface--font-family, Arial);
+    font-family: var(--interface--font-family, @sans-serif-fonts);
     font-weight: 500;
 }
