+++ packages/perspective-viewer/test/results/results.json
@@ -12,7 +12,7 @@
     "superstore.html/shows horizontal columns on small viewports.": "0b1b8977338c6d78536d21a05ae5827e",
     "superstore.html/setting a valid column should set it as input.": "e93df8b85b75225f02fc2ca6b1a789ba",
     "superstore.html/setting multiple column parameters should set input.": "086a24c19d4469773a041e0167fd5c55",
-    "superstore.html/computations should clear input column.": "b9dae91dbf8d15f97c59f74d52f741cf",
+    "superstore.html/computations should clear input column.": "733ff9d6e14a5abf96311137db343747",
     "superstore.html/saving a computed column should add it to inactive columns.": "e0efc884189a351ddb5e36d87c78c868",
     "superstore.html/aggregates by computed column.": "debc5a864c4dedf25a25eed7fb3b8ee6",
     "superstore.html/pivots by computed column.": "6522606fe66c2a8948ece023fb415354",
@@ -21,13 +21,13 @@
     "superstore.html/filters by computed column.": "929f9ce8dbd0fb057eb2c0f239eb6924",
     "blank.html/Handles reloading with a schema.": "b49d7dc76575e3e7d5814bf5cd7a2e88",
     "superstore.html/pivots by a column.": "fff4a8ca5b2bf809711f39549298e27e",
-    "superstore.html/click on add computed column button opens the UI.": "04d42dc7cf736a7551a8a2adc4d0a86f",
+    "superstore.html/click on add computed column button opens the UI.": "def086bfaef8cb8b211970a3b5392554",
     "superstore.html/click on close button closes the UI.": "534ea8c090373e6b8aeec01306a09ae4",
     "superstore.html/saving without parameters should show an error message.": "e6eda7825f6443708c1289d2ef3c6a8c",
     "superstore.html/doesn't leak tables.": "85b81ebf58de9d1d3cf667c2bdf9999c",
     "superstore.html/doesn't leak views when setting row pivots.": "2c416998c8c8263eea80c4106060858b",
     "superstore.html/doesn't leak views when setting filters.": "684adb3f82a3758e5e4ed6f237d416d3",
     "superstore.html/adds computed column via attribute": "f78b106d1a4c6ab6ca02e79869ee5ae9",
-    "superstore.html/saving without parameters should fail as button is disabled.": "04d42dc7cf736a7551a8a2adc4d0a86f",
+    "superstore.html/saving without parameters should fail as button is disabled.": "def086bfaef8cb8b211970a3b5392554",
     "superstore.html/saving a duplicate column should fail with error message.": "85b8669a4c9ba186ea914030c6b86a82"
 }
